requirements.txt
aiohttp>=3.9
websockets>=12.0
pyyaml>=6.0
fastapi>=0.110
uvicorn>=0.29
